# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PikPakGo API",
        "description": "PikPakGo Travel Marketplace API - Hotels, Vacation Rentals & More",
        "contact": {
            "name": "PikPakGo Support",
            "email": "reservations@pikpakgo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://pickpackgo.in-sourceit.com/api",
            "description": "Production Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/public/guest/session/update": {
            "post": {
                "tags": [
                    "Guest Session"
                ],
                "summary": "Update guest session with contact info",
                "operationId": "aceff3073ff856566b94aca5dcd73ce5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_id"
                                ],
                                "properties": {
                                    "session_id": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "first_name": {
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "country": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session updated"
                    },
                    "404": {
                        "description": "Session not found"
                    }
                }
            }
        },
        "/public/guest/session/{sessionId}": {
            "get": {
                "tags": [
                    "Guest Session"
                ],
                "summary": "Get guest session details",
                "operationId": "4b5b6020f137b9834c963300a50d7546",
                "parameters": [
                    {
                        "name": "sessionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session details"
                    },
                    "404": {
                        "description": "Session not found"
                    }
                }
            }
        },
        "/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get current user's notifications",
                "operationId": "b1274a2aaae8a4295a7e1925e8d8090f",
                "parameters": [
                    {
                        "name": "unread_only",
                        "in": "query",
                        "schema": {
                            "type": "boolean",
                            "example": false
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notifications list with unread count"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Delete all notifications for current user",
                "operationId": "83cb9e29aa2e091e93c6b2fc4a689921",
                "responses": {
                    "200": {
                        "description": "All deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/{id}/read": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark a notification as read",
                "operationId": "a28a3f705edbab5a620d9b567f97d739",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marked as read"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/read-all": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark all notifications as read",
                "operationId": "f437754b3bf6df932c41738206b4e101",
                "responses": {
                    "200": {
                        "description": "All marked as read"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notifications/{id}": {
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Delete a notification",
                "operationId": "0f28a0400610f81f01e2d25ce4e0edff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}